src/user/models.py
from django.db import models
from django.contrib.auth.models import AbstractUser, User
from django.forms import SelectDateWidget
from django.contrib import auth
from django.utils import timezone
from django.utils.text import slugify
from django_countries.fields import CountryField
from captcha.fields import ReCaptchaField
