src/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-countries==5.3.2
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-generate-secret-key==1.0.2
django-js-asset==1.1.0
django-recaptcha==1.4.0
django-wysiwyg==0.8.0
idna==2.7
mysqlclient==1.3.14
oauthlib==2.1.0
Pillow==5.3.0
pkg-resources==0.0.0
pycparser==2.19
pycrypto==2.6.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.12.0
urllib3==1.24.1
